Update tests to take into account "slower" code
A weird consequence of the way we do time is that measuring time
more often makes it go more slowly! Because we now measure time
before and after every top level draw, this causes a lot of tests
to appear to be very slow. This updates those tests to suppress
deadlines and health checks.

Note that this is not a real performance regression and is purely
an eccentricity of our test suite.