Improved calculation of useful intervals
This commit now adds all adjacent pairs of blocks as potentially
useful intervals. This allows us to improve the guarantees we make
about shrinking certain nested structures. See the updated test
for an example of that.

It also moves interval calculation to occur lazily, so we only
pay the cost of doing it when shrinking.