Remove vestigial zero-extension from tree lookup in cached_test_function
Logic for zero-extending input buffers was added in #996, but then
removed in favour of having callers add any necessary zeros manually.
However, this zero-extension step in the tree traversal was left behind.

In cases where the input buffer would overflow, but a zero-extended copy
of that buffer has already been tested and cached, this code would
result in one of two small problems: either the tree traversal continues
longer than necessary, or the cache returns a (non-overflow) result that
isn't accurate for the input buffer.