Prescreen capped bytes by masking, not by saturating
The buffer prescreening code was normalizing capped bytes with a
saturating min, but an actual test run normalizes them by masking
instead. This discrepancy would sometimes cause a novel buffer to be
rejected, or cause a non-novel buffer to be tested again.

This change makes the prescreen use a mask, so that it agrees with test
runs.