Assert that our byte masks are suitable for use as bounds
The values stored in 'capped' are actually byte masks, but there are two
places in ConjectureRunner that use them to compute range bounds or
range sizes instead.

This is fine, because the masks are always (2 ** n - 1) for some n,
which makes those computations valid.

This change adds assertions in those two places, to verify that the
masks match this expectation.