Improve subsequence drawing & reduce line lengths
This improves the proportions of booleans drawn for in the new
subsequence strategy, based on feedback provided in https://github.com/HypothesisWorks/hypothesis/pull/1533#pullrequestreview-149959035.
The ratio of elements masked to unmasked is now chosen uniformly, by
selecting a number of includes between the minimum and maximum number of
elements specified for the subsequence, and simply adding in the
remainder of elements as `False` values. This is an improvement on the
original implementation that chose 50-50 if elements where to be chosen
for each position after the minimum number of elements had already been
included.