Correctly restore global RNG state after using random_module()
The `random_module()` strategy sets global RNG state, and arranges for
`BuildContext` to restore the original state afterwards.

However, when `deterministic_PRNG` is nested inside `BuildContext`, this
causes the outer state-restoring hook to accidentally restore the
deterministic state, instead of the original state.

The solution is to reverse the nesting, so that each context manager
restores the intended state.