Removed usage of MD5 from calc_label_from_name
MD5 is cryptographically broken. Of course, this function doesn't really need cryptographic strength (it's slicing down to 8 bytes, so it hardly matters). However, when OpenSSL is used in FIPS mode (conforming to the US Government's FIPS-140 standard), MD5 is entirely disabled. While I generally have a pretty low opinion of FIPS, and I'm loath to ask projects to do actual work to support FIPS deployments, in this case I think the balance of equities is in favor:

- The match is tiny, and should not bring any ongoing maintenance burden
- FIPS's efforts to burn MD5 with fire are legitimately useful
- I'd expect the performance difference to be minimal, on my machine it's ~0.5 usecs per MD5 (of a 12 byte string) vs. ~0.6 usecs per SHA384 (same string)