Optimize character map generation
Optimize the hot loop over `range(0, sys.maxunicode)` (~1M iterations)
by reducing work when the current character has the same category as the
previous character.

This makes charmap generation about ~2x as fast (~0.80s => ~0.36s on my
system) and reduces the risk of the initial cache generation causing
a deadline failure in the test that triggers it.