Cache fewer inputs when fuzzing
The core idea here is that if we've previously seen some buffer that produces a particular failure, we only want to save the new buffer if it is shortlex smaller.

In practice this means that long fuzzing runs saturate and saves a lot of time clearing out irrelevant entries from the database later.