Doc fix: The database setting doesn't actually allow string values
This docstring suggests that callers can pass a database path string or
the special string ":memory:", but those arguments are not actually
permitted by the implementation.

It seems that an early draft of #1196 did allow string arguments (to
match the now-removed database_file setting). The feature was then
removed after review feedback, but unfortunately the docstring was not
updated.

In-memory DBs and custom DB paths can still be specified by explicitly
instantiating a subclass of ExampleDatabase, or by calling
ExampleDatabase with a string argument (currently under-documented?).