Add HealthCheck.function_scoped_fixture to suppress fixture warnings
The Hypothesis pytest plugin emits a warning when it sees `given`
applied to a test function that also takes pytest function-scoped
fixtures. Function-scoped fixtures are only run once per function, not
per example, which is usually unexpected and can cause subtle problems.

In rare cases, once-per-function behaviour is specifically intended.
Currently those cases can be accommodated by suppressing the warning,
but that will no longer be the case when a future release changes the
warning into an error by default.

This change therefore introduces the HealthCheck.function_scoped_fixture
flag, which provides an alternate way to suppress those checks for one
or more affected test functions. Tests that use this flag to knowingly
suppress the warning will continue to work when the warning becomes a
health check error.