Report stats in HTML
A tiny always-on integration with pytest-html, where reporting our statistics summary doesn't take up space like it would on the command-line.

Note that this is untested, because there's only one very simple line of code, and it would be painful to test even if we set up another virtualenv with pytest-html installed.