Remove `COMPLIANT_XP` flag for `tests/array_api`
This flag is only used to skip testing __array_namespace__() for arrays created
by the mocked Array API namespace. This proves inconvenient when parametrizing
these tests for multiple array libraries, as you have to keep track of this
variable.

Testing whether the generated arrays have this method is in an awkard place
between:

* Testing whether generated arrays are indeed arrays
* Testing whether the array module has implemented this method

This commit therefore opts to favour simplicity, as the utility of testing
__array_namespace__() is in a grey area anywho.