Wait for mypy process end
The former code launched the subprocess via .Popen, read its output, but never
waited for the process to end. This could result in no/partial output (I'm
surprised it seemed to have worked properly!), and also results in a
ResourceWarning with Python/pytest being configured correctly, as a
still-running subprocess object gets GC'ed.