Allow registration of tuples
Previously, if `tuple` was registered, the registered strategy would be
ignored. Now it is actually used. Note that `tuple[...]` cannot be
registered because genericized types are not allowed to be registered in
general, so a user must register `tuple` with a function that takes the
type args and returns a strategy appropriate to the particular
genericized type at hand.