Harmonize cache implementations with documentation.
There was a potential bug here, in that the entry score was not updated
on access for concrete cache implementations if they followed the
documentated behaviour of on_access. The only in-tree implementation
worked fine, as it did not.

Also makes the documentation more specific in a couple of places.