Use tmp_path in ghostwriter tests
Before creating the temporary functions, moves the local directory to a
clean temporary directory; this should guarantee that the files to be
created don't exist, and handles cleanup automatically.

This also works when the local directory that the tests are executed
from is not writable, but the temporary directory is.