Restore compatibility with Python 3.9's legacy LL(1) parser yet again
As a follow-up to 422425f3a1f322322f5783c4aa6a10ca54cfd5dc, the legacy
LL(1) parser also parses `with (expr1, expr2):` as simple
with-statements, with a single tuple as the context manager expression.
Such an expression can easily arise accidentally due to the use of an
auto-formatter. Guard against this by rewriting these statements as
nested with-statements and adding an appropriate comment about LL(1)
parsability.